+++ src/Components/User Account/Edit.js
@@ -1,4 +1,4 @@
-import Header from "../Header";
+import Header from "../Header/Header";
 import EditForm from "./EditForm";
 
 export default function Edit() {
